aiot_fresh/requirements.txt
flask
paho-mqtt
firebase-admin
requests
gunicorn
python-dotenv
